Build altlist-implementations conditionally
