+++ tests/extendrtests/man/new_usize.Rd
@@ -1,0 +1,16 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/extendr-wrappers.R
+\name{new_usize}
+\alias{new_usize}
+\title{Create an ALTLIST usize vector}
+\usage{
+new_usize(robj)
+}
+\arguments{
+\item{robj}{an integer vector
+
+The object is \verb{Vec<Option<usize>>} represented as an ALTLIST}
+}
+\description{
+Create an ALTLIST usize vector
+}
