+++ tests/extendrtests/man/tst_altinteger.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/extendr-wrappers.R
+\name{tst_altinteger}
+\alias{tst_altinteger}
+\title{Test ALTINTEGER support}
+\usage{
+tst_altinteger()
+}
+\description{
+Test ALTINTEGER support
+}
