+++ tests/extendrtests/man/tst_altstring.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/extendr-wrappers.R
+\name{tst_altstring}
+\alias{tst_altstring}
+\title{Test ALTSTRING representation}
+\usage{
+tst_altstring()
+}
+\description{
+Test ALTSTRING representation
+}
